Create connections.pl
This file is used to connecting to local MySQL server.  It was placed in a separate file so that I can avoid duplicate connection code in multiple files.  Plus one change with this file affects the entire application which is helpful when an application moves from the development to production environment.